+++ pkg/golang/sum/parse_test.go
@@ -1,4 +1,4 @@
-package mod
+package sum
 
 import (
 	"os"
